# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "alex production",
        "description": "Documentation for alex",
        "contact": {
            "email": "alex@gmail"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/founder": {
            "get": {
                "tags": [
                    "Founders"
                ],
                "summary": "Get list of Founder",
                "description": "Returns list of Founder",
                "operationId": "indexFounder",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Founder"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Founders"
                ],
                "summary": "Insert a new Founder",
                "description": "Insert a new Founder",
                "operationId": "storeFounder",
                "requestBody": {
                    "description": "Founder to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Founder"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Founder created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Founder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/founder/{founder_id}": {
            "get": {
                "tags": [
                    "Founders"
                ],
                "summary": "Show a Founder from his Id",
                "description": "Show a Founder from his Id",
                "operationId": "showFounder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Founder--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Founder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Founder not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Founders"
                ],
                "summary": "Delete a Founder",
                "description": "Delete a Founder",
                "operationId": "destroyFounder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Founder--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Founder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Founder not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Founders"
                ],
                "summary": "Update an existing Founder",
                "description": "Update an existing Founder",
                "operationId": "updateFounder",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Founder--id"
                    }
                ],
                "requestBody": {
                    "description": "Founder to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Founder"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Founder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Founder not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Founder": {
                "title": "Founder",
                "description": "Founder model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Founder",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Founder",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Founder",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Founder",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "Founder--id": {
                "name": "Founder_id",
                "in": "path",
                "description": "Id of Founder",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        }
    }
}